refactor: move encoding code into token.js
