+++ packages/client/src/lib.js
@@ -19,13 +19,10 @@ import pRetry from 'p-retry'
 import { TreewalkCarSplitter } from 'carbites/treewalk'
 import { pack } from 'ipfs-car/pack'
 import { CID } from 'multiformats/cid'
-import * as Block from 'multiformats/block'
-import { sha256 } from 'multiformats/hashes/sha2'
-import * as dagCbor from '@ipld/dag-cbor'
-import { BlockstoreCarReader } from './blockstore.js'
 import * as Token from './token.js'
 import { fetch, File, Blob, FormData, Blockstore } from './platform.js'
 import { toGatewayURL } from './gateway.js'
+import { BlockstoreCarReader } from './bs-car-reader.js'
 
 const MAX_STORE_RETRIES = 5
 const MAX_CONCURRENT_UPLOADS = 3
@@ -144,33 +141,32 @@ class NFTStorage {
         ? await TreewalkCarSplitter.fromBlob(car, targetSize, { decoders })
         : new TreewalkCarSplitter(car, targetSize, { decoders })
 
-    const upload = transform(
-      MAX_CONCURRENT_UPLOADS,
-      async function (/** @type {AsyncIterable<Uint8Array>} */ car) {
-        const carParts = []
-        for await (const part of car) {
-          carParts.push(part)
-        }
-        const carFile = new Blob(carParts, { type: 'application/car' })
-        const cid = await pRetry(
-          async () => {
-            const response = await fetch(url.toString(), {
-              method: 'POST',
-              headers: NFTStorage.auth(token),
-              body: carFile,
-            })
-            const result = await response.json()
-            if (!result.ok) {
-              throw new Error(result.error.message)
-            }
-            return result.value.cid
-          },
-          { retries: maxRetries == null ? MAX_STORE_RETRIES : maxRetries }
-        )
-        onStoredChunk && onStoredChunk(carFile.size)
-        return cid
+    const upload = transform(MAX_CONCURRENT_UPLOADS, async function(
+      /** @type {AsyncIterable<Uint8Array>} */ car
+    ) {
+      const carParts = []
+      for await (const part of car) {
+        carParts.push(part)
       }
-    )
+      const carFile = new Blob(carParts, { type: 'application/car' })
+      const cid = await pRetry(
+        async () => {
+          const response = await fetch(url.toString(), {
+            method: 'POST',
+            headers: NFTStorage.auth(token),
+            body: carFile,
+          })
+          const result = await response.json()
+          if (!result.ok) {
+            throw new Error(result.error.message)
+          }
+          return result.value.cid
+        },
+        { retries: maxRetries == null ? MAX_STORE_RETRIES : maxRetries }
+      )
+      onStoredChunk && onStoredChunk(carFile.size)
+      return cid
+    })
 
     let root
     for await (const cid of upload(splitter.cars())) {
@@ -238,61 +234,16 @@ class NFTStorage {
     validateERC1155(input)
     const blockstore = new Blockstore()
     try {
-      const [blobs, meta] = Token.encode(input)
-      /** @type {import('./lib/interface').Encoded<T, [[Blob, URL]]>} */
-      const data = JSON.parse(JSON.stringify(meta))
-      /** @type {import('./lib/interface').Encoded<T, [[Blob, CID]]>} */
-      const dag = JSON.parse(JSON.stringify(meta))
-
-      for (const [dotPath, blob] of blobs.entries()) {
-        /** @type {string|undefined} */
-        // @ts-ignore blob may be a File!
-        const name = blob.name || 'blob'
-        const { root: cid } = await pack({
-          // @ts-ignore
-          input: [{ path: name, content: blob.stream() }],
-          blockstore,
-          wrapWithDirectory: true,
-        })
-
-        const href = new URL(`ipfs://${cid}/${name}`)
-        const path = dotPath.split('.')
-        setIn(data, path, href)
-        setIn(dag, path, cid)
-      }
-
-      const { root: metadataJsonCid } = await pack({
-        // @ts-ignore
-        input: [
-          {
-            path: 'metadata.json',
-            content: new Blob([JSON.stringify(data)]).stream(),
-          },
-        ],
-        blockstore,
-        wrapWithDirectory: false,
-      })
-
-      const block = await Block.encode({
-        value: {
-          ...dag,
-          'metadata.json': metadataJsonCid,
-          type: 'nft',
-        },
-        codec: dagCbor,
-        hasher: sha256,
-      })
-      await blockstore.put(block.cid, block.bytes)
-
-      onRootCidReady && onRootCidReady(block.cid.toString())
-      const car = new BlockstoreCarReader(1, [block.cid], blockstore)
-      await NFTStorage.storeCar(service, car, { onStoredChunk, maxRetries })
-
-      return new Token.Token(
-        block.cid.toString(),
-        `ipfs://${block.cid}/metadata.json`,
-        data
+      const token = await Token.encode(input, blockstore)
+      onRootCidReady && onRootCidReady(token.ipnft)
+      const car = new BlockstoreCarReader(
+        1,
+        [CID.parse(token.ipnft)],
+        blockstore
       )
+      await NFTStorage.storeCar(service, car, { onStoredChunk, maxRetries })
+      // @ts-ignore
+      return token
     } finally {
       await blockstore.close()
     }
@@ -568,8 +519,8 @@ For more context please see ERC-721 specification https://eips.ethereum.org/EIPS
  * @param {Deal[]} deals
  * @returns {Deal[]}
  */
-const decodeDeals = (deals) =>
-  deals.map((deal) => {
+const decodeDeals = deals =>
+  deals.map(deal => {
     const { dealActivation, dealExpiration, lastChanged } = {
       dealExpiration: null,
       dealActivation: null,
@@ -588,33 +539,6 @@ const decodeDeals = (deals) =>
  * @param {Pin} pin
  * @returns {Pin}
  */
-const decodePin = (pin) => ({ ...pin, created: new Date(pin.created) })
-
-/**
- * Sets a given `value` at the given `path` on a passed `object`.
- *
- * @example
- * ```js
- * const obj = { a: { b: { c: 1 }}}
- * setIn(obj, ['a', 'b', 'c'], 5)
- * obj.a.b.c //> 5
- * ```
- *
- * @template V
- * @param {any} object
- * @param {string[]} path
- * @param {V} value
- */
-const setIn = (object, path, value) => {
-  const n = path.length - 1
-  let target = object
-  for (let [index, key] of path.entries()) {
-    if (index === n) {
-      target[key] = value
-    } else {
-      target = target[key]
-    }
-  }
-}
+const decodePin = pin => ({ ...pin, created: new Date(pin.created) })
 
 export { NFTStorage, File, Blob, FormData, toGatewayURL, Token }
